+++ extensions/sql-database-projects/src/templates/templates.ts
@@ -7,7 +7,7 @@ import * as path from "path";
 import { promises as fs } from "fs";
 import { ItemType } from "sqldbproj";
 import * as constants from "../common/constants";
-import { getMicrosoftBuildSqlVersion } from "../tools/netcoreTool";
+import { getMicrosoftBuildSqlVersion, resolveNugetVersion } from "../tools/netcoreTool";
 
 export let newSqlProjectTemplate: string;
 export let newSdkSqlProjectTemplate: string;
@@ -82,7 +82,15 @@ export async function loadTemplates(templateFolderPath: string) {
         Promise.resolve(
             (newSdkSqlProjectTemplate = macroExpansion(
                 await loadTemplate(templateFolderPath, "newSdkSqlProjectTemplate.xml"),
-                new Map([["MICROSOFT_BUILD_SQL_VERSION", getMicrosoftBuildSqlVersion()]]),
+                new Map([
+                    [
+                        "MICROSOFT_BUILD_SQL_VERSION",
+                        await resolveNugetVersion(
+                            "Microsoft.Build.Sql",
+                            getMicrosoftBuildSqlVersion(),
+                        ),
+                    ],
+                ]),
             )),
         ),
         loadObjectTypeInfo(